+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Change Log
 
 ### Bug Fixes
 * Corrigir uso de PageRequest na API v3 (PL-3704)
+* Corrige busca de veículo por placa
 
 <a name="v3.5.1"></a>
 ## Version [v3.5.1](https://github.com/luizfp/PrologWebService/compare/v3.5.0...v3.5.1) (2021-06-29)